refactor render_item based classes
All render_item based classes are moved into the own headers
