+++ include/litehtml/render_item.h
@@ -273,8 +273,6 @@ namespace litehtml
             ri->parent(shared_from_this());
         }
 
-        int render(int x, int y, const containing_block_context& containing_block_size, formatting_context* fmt_ctx, bool second_pass = false);
-
 		bool is_root() const
 		{
 			return m_parent.expired();
@@ -305,6 +303,7 @@ namespace litehtml
             return !(m_skip || src_el()->css().get_display() == display_none || src_el()->css().get_visibility() != visibility_visible);
         }
 
+		int render(int x, int y, const containing_block_context& containing_block_size, formatting_context* fmt_ctx, bool second_pass = false);
         int calc_width(int defVal, int containing_block_width) const;
         bool get_predefined_height(int& p_height, int containing_block_height) const;
         void apply_relative_shift(const containing_block_context &containing_block_size);
@@ -348,225 +347,6 @@ namespace litehtml
          */
         void get_rendering_boxes( position::vector& redraw_boxes);
 	};
-
-    class render_item_block : public render_item
-    {
-    protected:
-		/**
-		 * Render block content.
-		 *
-		 * @param x - horizontal position of the content
-		 * @param y - vertical position of the content
-		 * @param second_pass - true is this is the second pass.
-		 * @param ret_width - input minimal width.
-		 * @param self_size - defines calculated size of block
-		 * @return return value is the minimal width of the content in block. Must be greater or equal to ret_width parameter
-		 */
-        virtual int _render_content(int x, int y, bool second_pass, const containing_block_context &self_size, formatting_context* fmt_ctx) {return 0;}
-		int _render(int x, int y, const containing_block_context &containing_block_size, formatting_context* fmt_ctx, bool second_pass) override;
-        int place_float(const std::shared_ptr<render_item> &el, int top, const containing_block_context &self_size, formatting_context* fmt_ctx);
-		virtual void fix_line_width(element_float flt,
-									const containing_block_context &containing_block_size, formatting_context* fmt_ctx)
-		{}
-
-    public:
-        explicit render_item_block(std::shared_ptr<element>  src_el) : render_item(std::move(src_el))
-        {}
-
-        std::shared_ptr<render_item> clone() override
-        {
-            return std::make_shared<render_item_block>(src_el());
-        }
-        std::shared_ptr<render_item> init() override;
-    };
-
-    /**
-     * In a block formatting context, boxes are laid out one after the other, vertically, beginning at the top of a
-     * containing block.
-     * https://www.w3.org/TR/CSS22/visuren.html#block-formatting
-     */
-    class render_item_block_context : public render_item_block
-    {
-    protected:
-        int _render_content(int x, int y, bool second_pass, const containing_block_context &self_size, formatting_context* fmt_ctx) override;
-
-    public:
-        explicit render_item_block_context(std::shared_ptr<element>  src_el) : render_item_block(std::move(src_el))
-        {}
-
-        std::shared_ptr<render_item> clone() override
-        {
-            return std::make_shared<render_item_block_context>(src_el());
-        }
-    };
-
-    /**
-     * An inline formatting context is established by a block container box that contains no block-level boxes.
-     * https://www.w3.org/TR/CSS22/visuren.html#inline-formatting
-     */
-    class render_item_inline_context : public render_item_block
-    {
-		/**
-		 *	Structure contains elements with display: inline
-		 *	members:
-		 *	- element: 		render_item with display: inline
-		 *	- boxes: 		rectangles represented inline element content. There are can be many boxes if content
-		 *			 is split into some lines
-		 *	- start_box:	the start position of currently calculated box
-		 */
-		struct inlines_item
-		{
-			std::shared_ptr<render_item> element;
-			position::vector boxes;
-			position	start_box;
-
-			explicit inlines_item(const std::shared_ptr<render_item>& el) : element(el) {}
-		};
-    protected:
-        std::vector<std::unique_ptr<litehtml::line_box> > m_line_boxes;
-		int m_max_line_width;
-
-        int _render_content(int x, int y, bool second_pass, const containing_block_context &self_size, formatting_context* fmt_ctx) override;
-        void fix_line_width(element_float flt,
-							const containing_block_context &self_size, formatting_context* fmt_ctx) override;
-
-        std::list<std::unique_ptr<line_box_item> > finish_last_box(bool end_of_render, const containing_block_context &self_size);
-        void place_inline(std::unique_ptr<line_box_item> item, const containing_block_context &self_size, formatting_context* fmt_ctx);
-        int new_box(const std::unique_ptr<line_box_item>& el, line_context& line_ctx, const containing_block_context &self_size, formatting_context* fmt_ctx);
-        void apply_vertical_align() override;
-    public:
-        explicit render_item_inline_context(std::shared_ptr<element>  src_el) : render_item_block(std::move(src_el)), m_max_line_width(0)
-        {}
-
-        std::shared_ptr<render_item> clone() override
-        {
-            return std::make_shared<render_item_inline_context>(src_el());
-        }
-
-        int get_base_line() override;
-    };
-
-    class render_item_table : public render_item
-    {
-    protected:
-        // data for table rendering
-        std::unique_ptr<table_grid>	m_grid;
-        int						    m_border_spacing_x;
-        int						    m_border_spacing_y;
-
-		int _render(int x, int y, const containing_block_context &containing_block_size, formatting_context* fmt_ctx, bool second_pass) override;
-
-    public:
-        explicit render_item_table(std::shared_ptr<element>  src_el);
-
-        std::shared_ptr<render_item> clone() override
-        {
-            return std::make_shared<render_item_table>(src_el());
-        }
-        void draw_children(uint_ptr hdc, int x, int y, const position* clip, draw_flag flag, int zindex) override;
-        int get_draw_vertical_offset() override;
-        std::shared_ptr<render_item> init() override;
-    };
-
-	class render_item_table_part : public render_item
-	{
-	public:
-		explicit render_item_table_part(std::shared_ptr<element>  src_el) : render_item(std::move(src_el))
-		{}
-
-		std::shared_ptr<render_item> clone() override
-		{
-			return std::make_shared<render_item_table_part>(src_el());
-		}
-	};
-
-	class render_item_table_row : public render_item
-	{
-	public:
-		explicit render_item_table_row(std::shared_ptr<element>  src_el) : render_item(std::move(src_el))
-		{}
-
-		std::shared_ptr<render_item> clone() override
-		{
-			return std::make_shared<render_item_table_row>(src_el());
-		}
-		void get_inline_boxes( position::vector& boxes ) const override;
-	};
-
-	class render_item_inline : public render_item
-    {
-    protected:
-		position::vector m_boxes;
-
-    public:
-        explicit render_item_inline(std::shared_ptr<element>  src_el) : render_item(std::move(src_el))
-        {}
-
-		void get_inline_boxes( position::vector& boxes ) const override { boxes = m_boxes; }
-		void set_inline_boxes( position::vector& boxes ) override { m_boxes = boxes; }
-		void add_inline_box( const position& box ) override { m_boxes.emplace_back(box); };
-		void clear_inline_boxes() override { m_boxes.clear(); }
-		int get_base_line() override { return src_el()->css().get_font_metrics().base_line(); }
-
-		std::shared_ptr<render_item> clone() override
-        {
-            return std::make_shared<render_item_inline>(src_el());
-        }
-    };
-
-    class render_item_image : public render_item
-    {
-    protected:
-        int calc_max_height(int image_height, int containing_block_height);
-		int _render(int x, int y, const containing_block_context &containing_block_size, formatting_context* fmt_ctx, bool second_pass) override;
-
-    public:
-        explicit render_item_image(std::shared_ptr<element>  src_el) : render_item(std::move(src_el))
-        {}
-
-        std::shared_ptr<render_item> clone() override
-        {
-            return std::make_shared<render_item_image>(src_el());
-        }
-    };
-
-    class render_item_flex : public render_item_block
-    {
-        struct flex_item
-        {
-            std::shared_ptr<render_item> el;
-            int base_size;
-            int main_size;
-            int min_width;
-            int max_width;
-            int line;
-
-            explicit flex_item(std::shared_ptr<render_item>  _el) :
-                el(std::move(_el)),
-                min_width(0),
-                max_width(0),
-                line(0),
-                base_size(0),
-                main_size(0)
-            {}
-        };
-    protected:
-        std::list<std::unique_ptr<flex_item>>   m_flex_items;
-
-        int _render_content(int x, int y, bool second_pass, const containing_block_context &self_size, formatting_context* fmt_ctx) override;
-
-    public:
-        explicit render_item_flex(std::shared_ptr<element>  src_el) : render_item_block(std::move(src_el))
-        {}
-
-        std::shared_ptr<render_item> clone() override
-        {
-            return std::make_shared<render_item_flex>(src_el());
-        }
-        void draw_children(uint_ptr hdc, int x, int y, const position* clip, draw_flag flag, int zindex) override;
-        std::shared_ptr<render_item> init() override;
-    };
-
 }
 
 #endif //LH_RENDER_ITEM_H
