+++ src/render_block.cpp
@@ -1,5 +1,7 @@
 #include "html.h"
-#include "render_item.h"
+#include "render_block.h"
+#include "render_inline_context.h"
+#include "render_block_context.h"
 #include "document.h"
 
 int litehtml::render_item_block::place_float(const std::shared_ptr<render_item> &el, int top, const containing_block_context &self_size, formatting_context* fmt_ctx)
