+++ src/render_flex.cpp
@@ -1,6 +1,6 @@
 #include "html.h"
-#include "render_item.h"
 #include "types.h"
+#include "render_flex.h"
 
 int litehtml::render_item_flex::_render_content(int x, int y, bool second_pass, const containing_block_context &self_size, formatting_context* fmt_ctx)
 {
