--- src/render_inline.cpp
@@ -1,3 +1,0 @@
-#include "html.h"
-#include "render_item.h"
-
