+++ src/render_inline_context.cpp
@@ -1,5 +1,5 @@
 #include "html.h"
-#include "render_item.h"
+#include "render_inline_context.h"
 #include "document.h"
 #include "iterators.h"
