+++ src/render_table.cpp
@@ -1,5 +1,5 @@
 #include "html.h"
-#include "render_item.h"
+#include "render_table.h"
 #include "document.h"
 #include "iterators.h"
 
@@ -461,4 +461,22 @@ int litehtml::render_item_table::get_draw_vertical_offset()
         return m_grid->captions_height();
     }
     return 0;
+}
+
+void litehtml::render_item_table_row::get_inline_boxes( position::vector& boxes ) const
+{
+	position pos;
+	for(auto& el : m_children)
+	{
+		if(el->src_el()->css().get_display() == display_table_cell)
+		{
+			pos.x		= el->left() + el->margin_left();
+			pos.y		= el->top() - m_padding.top - m_borders.top;
+
+			pos.width	= el->right() - pos.x - el->margin_right() - el->margin_left();
+			pos.height	= el->height() + m_padding.top + m_padding.bottom + m_borders.top + m_borders.bottom;
+
+			boxes.push_back(pos);
+		}
+	}
 }
